Everything is component now
